backend/requirements.txt
asgiref
Django
djangorestframework
django-cors-headers
django-taggit
pytz
ruff
sqlparse
